moved ResourceManager
